Extract shared mock setup for useAIWorkflowApplications tests
